AI-Market-Aggregator-New/requirements.txt
# requirements.txt - Production dependencies for AI Market Intelligence Aggregator

# Core dependencies (required)
requests>=2.31.0        # HTTP client with security fixes
feedparser>=6.0.10      # Robust RSS/Atom parsing (replaces vulnerable regex)
markdown>=3.4.0         # Professional markdown-to-HTML conversion for email reports
PyYAML>=6.0             # For parsing the new ai_config.yml

# Future enhancements (commented out until needed)
# aiohttp>=3.9.0        # For async HTTP requests (Phase 2 performance)
# boto3>=1.34.0         # AWS SDK for Lambda migration (Phase 3)
# jinja2>=3.1.0         # HTML templating engine (Phase 2 enhancement)

# Development tools (optional)
python-dotenv>=0.21.0   # For loading .env files during local development
pytest>=7.4.0         # Testing framework
# black>=23.0.0         # Code formatting
# flake8>=6.0.0         # Linting
